#24 updated navbar
