+++ clientapp/src/modules/navbar/components/NavBar/index.js
@@ -1,100 +1,104 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import withStyles from '@mui/styles/withStyles';
-import { NavLink } from 'react-router-dom'
-import { Avatar, List, ListItem, ListItemText } from '@mui/material'
+import withStyles from '@mui/styles/withStyles'
+import { Link } from 'react-router-dom'
+import {
+  Avatar,
+  AppBar,
+  Toolbar,
+  Typography,
+  Container,
+  Box,
+  Button,
+  IconButton,
+  Menu,
+  MenuItem,
+} from '@mui/material'
 
 import styles from './styles'
 import routes from '@movie/routes'
+import NavbarTabs from '../NavbarTabs'
 
-const NavBar = ({ classes, isLogined, onClickLogout, fullName, avatar }) => (
-  <nav className={classes.navbarBlock}>
-    <List
-      component='nav'
-      className={`${classes.root} ${classes.profileSectionsBlock}`}
-      aria-label='mailbox folders'
-    >
-      <NavLink
-        to={routes.root}
-        className={classes.navBarLink}
-        hover='true'
-        activeClassName={classes.activeNavBarLink}
-        exact
-      >
-        <ListItem>
-          <ListItemText className={classes.listItemText} primary='Home' />
-        </ListItem>
-      </NavLink>
-      {
-        isLogined
-          ? null
-          : (
-          <NavLink
-            to={routes.login}
-            className={classes.navBarLink}
-            hover='true'
-            activeClassName={classes.activeNavBarLink}
-          >
-            <ListItem>
-              <ListItemText className={classes.listItemText} primary='Login' />
-            </ListItem>
-          </NavLink>
-          )
-      }
-      {
-        isLogined
-          ? null
-          : (
-          <NavLink
-            to={routes.register}
-            className={classes.navBarLink}
-            hover='true'
-            activeClassName={classes.activeNavBarLink}
-          >
-            <ListItem>
-              <ListItemText className={classes.listItemText} primary='Register' />
-            </ListItem>
-          </NavLink>
-          )
-      }
-      {
-        !isLogined
-          ? null
-          : (
-          <NavLink
-            to={routes.profile}
-            className={classes.navBarLink}
-            hover='true'
-            activeClassName={classes.activeNavBarLink}
-          >
-            <ListItem>
-              <Avatar src={avatar} className={classes.avatarBlock} />
-              <ListItemText className={classes.listItemText} primary={fullName} />
-            </ListItem>
-          </NavLink>
-          )
-      }
-      {
-        !isLogined
-          ? null
-          : (
-          <ListItem>
-            <button className={classes.navBarButton} onClick={onClickLogout}>
-              Logout
-            </button>
-          </ListItem>
-          )
-      }
-    </List>
-  </nav>
+const settings = [
+  {
+    route: routes.profile,
+    label: 'Profile',
+  },
+]
+
+const MuiNavbar = ({ 
+  classes,
+  isLogined,
+  onClickLogout,
+  avatar,
+  anchorUser,
+  handleOpenUserMenu,
+  handleCloseUserMenu,
+}) => (
+  <AppBar position='static'>
+    <Container maxWidth='lg'>
+      <Toolbar disableGutters>
+        <Box sx={{ flexGrow: 1, display: 'flex' }}>
+          <NavbarTabs />
+        </Box>
+        <Box sx={{ flexGrow: 0 }}>
+          {
+            isLogined
+            ?
+            <>
+              <IconButton onClick={handleOpenUserMenu}>
+                <Avatar src={avatar} className={classes.avatarBlock} />
+              </IconButton>
+              <Menu 
+                sx={{ mt: '45px'}}
+                anchorEl={anchorUser}
+                open={Boolean(anchorUser)}
+                onClose={handleCloseUserMenu}
+                anchorOrigin={{
+                  vertical: 'top',
+                  horizontal: 'right',
+                }}
+                keepMounted
+                transformOrigin={{
+                  vertical: 'top',
+                  horizontal: 'right',
+                }}
+              >
+              {
+                settings.map(setting => (
+                  <MenuItem key={setting.label} component={Link} to={setting.route} onClick={handleCloseUserMenu}>
+                    <Typography textAlign='center'>{setting.label}</Typography>
+                  </MenuItem>
+                ))
+              }
+                <MenuItem onClick={() => {
+                  onClickLogout()
+                  handleCloseUserMenu()
+                  }}>
+                  <Typography textAlign='center'>Logout</Typography>
+                </MenuItem>
+              </Menu>
+            </>
+            : 
+            <Button component={Link} to={routes.login} sx={{ color: 'white' }}>
+              Login
+            </Button>
+          }
+          
+        </Box>
+      </Toolbar>
+    </Container>
+  </AppBar>
 )
 
-NavBar.propTypes = {
+MuiNavbar.propTypes = {
   classes: PropTypes.object.isRequired,
   isLogined: PropTypes.bool,
-  onClickLogout: PropTypes.func.isRequired,
-  fullName: PropTypes.string,
   avatar: PropTypes.string,
+  onClickLogout: PropTypes.func.isRequired,
+  anchorUser: PropTypes.object,
+  handleOpenUserMenu: PropTypes.func.isRequired,
+  handleCloseUserMenu: PropTypes.func.isRequired,
 }
 
-export default withStyles(styles)(NavBar)
+export default withStyles(styles)(MuiNavbar)
