+++ clientapp/src/modules/navbar/components/NavbarTabs/styles.js
@@ -1,0 +1,4 @@
+const styles = {
+}
+
+export default styles
