--- clientapp/src/modules/navbar/components/muiNavbar/styles.js
@@ -1,4 +1,0 @@
-const styles = {
-}
-
-export default styles
