+++ clientapp/src/modules/navbar/containers/NavBarContainer.js
@@ -1,18 +1,15 @@
 import React, { useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useHistory } from 'react-router-dom'
-import { Container  } from '@mui/material'
 
-import NavBar from '../components/NavBar'
 import { logoutRequest } from '../../account/actions'
-import { getFullName, getUserAvatar, getIsLoginSucceeded } from '../../account/selectors'
+import { getUserAvatar, getIsLoginSucceeded } from '../../account/selectors'
 import routes from '@movie/routes'
-import MuiNavbar from '../components/muiNavbar'
+import Navbar from '../components/Navbar'
 
 const NavBarContainer = () => {
   const dispatch = useDispatch()
   const isLogined = useSelector(getIsLoginSucceeded)
-  const fullName = useSelector(getFullName)
   const avatar = useSelector(getUserAvatar)
   const history = useHistory()
 
@@ -30,25 +27,14 @@ const NavBarContainer = () => {
   }
 
   return (
-    <>
-      <Container maxWidth="lg">
-        <NavBar
-          isLogined={isLogined}
-          fullName={fullName}
-          avatar={avatar}
-          onClickLogout={onClickLogout}
-        />
-      </Container>
-      <MuiNavbar 
-        isLogined={isLogined}
-        fullName={fullName}
-        avatar={avatar}
-        onClickLogout={onClickLogout}
-        anchorUser={anchorUser}
-        handleOpenUserMenu={handleOpenUserMenu}
-        handleCloseUserMenu={handleCloseUserMenu}
-      />
-    </>
+    <Navbar 
+      isLogined={isLogined}
+      avatar={avatar}
+      onClickLogout={onClickLogout}
+      anchorUser={anchorUser}
+      handleOpenUserMenu={handleOpenUserMenu}
+      handleCloseUserMenu={handleCloseUserMenu}
+    />
   )
 }
